chore(style): separate files for cmd and query fns
